mcp-server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sympy>=1.9
pydantic>=1.8.2
requests>=2.25.1
openai>=1.0.0
httpx>=0.24.0
PyYAML>=6.0